Restructured files and made CliArgs its own proper library target
