+++ GridKit/Utilities/CliArgs/CMakeLists.txt
@@ -1,0 +1,10 @@
+gridkit_add_library(utilities_cli_args
+  SOURCES
+    CliArgs.cpp
+  HEADERS
+    CliArgs.hpp
+  LINK_LIBRARIES
+    PUBLIC GridKit::utilities_logger
+  )
+
+target_link_libraries(Utilities INTERFACE GridKit::utilities_cli_args)
