+++ GridKit/Utilities/CliOptions/CMakeLists.txt
@@ -1,9 +1,7 @@
 gridkit_add_library(utilities_cli_options
   SOURCES
-    CliArgs.cpp
     CliOptions.cpp
   HEADERS
-    CliArgs.hpp
     CliOptions.hpp
   LINK_LIBRARIES
     PUBLIC GridKit::utilities_logger
